update api env
