+++ frontend/src/lib/WorkCard.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
-	import WorkTag from '$lib/WorkTag.svelte';
 	import DisplayText from '$lib/DisplayText.svelte';
+	import WorkTag from '$lib/WorkTag.svelte';
 	import WorkThumbnail from '$lib/WorkThumbnail.svelte';
 	import { m } from '$lib/paraglide/messages.js';
-	import { getDisplayText } from '$lib/api';
+	import { getDisplayText } from '$lib/ui.js';
 	import { Status, type components } from './schema';
 
 	interface Props {
