+++ frontend/src/routes/(auth)/login/+page.server.ts
@@ -1,7 +1,7 @@
 import { fail, redirect, type Actions } from '@sveltejs/kit';
 import type { PageServerLoad } from './$types';
 import client from '$lib/api.server';
-import { forwardCookies } from '$lib/api';
+import { forwardCookies } from '$lib/api.server';
 import { m } from '$lib/paraglide/messages';
 
 export const load: PageServerLoad = async ({ cookies, fetch, locals, url }) => {
