+++ frontend/src/routes/(auth)/logout/+page.server.ts
@@ -1,4 +1,4 @@
-import client from '$lib/api';
+import { rawClient } from '$lib/api.server';
 import { redirect } from '@sveltejs/kit';
 import type { Actions } from './$types';
 
@@ -6,7 +6,7 @@ export const actions: Actions = {
 	default: async ({ cookies, fetch, locals }) => {
 		if (!locals.user) redirect(303, '/');
 
-		const { error } = await client.POST('/api/auth/logout', { fetch });
+		const { error } = await rawClient.POST('/api/auth/logout', { fetch });
 
 		if (!error) {
 			cookies.delete('csrftoken', { path: '/' });
