+++ frontend/src/routes/(auth)/reset_password/+page.server.ts
@@ -1,7 +1,8 @@
 import { fail, type Actions } from '@sveltejs/kit';
 import type { PageServerLoad } from './$types';
 import client from '$lib/api.server';
-import clientRaw, { forwardCookies } from '$lib/api';
+import clientRaw from '$lib/api';
+import { forwardCookies } from '$lib/api.server';
 import { m } from '$lib/paraglide/messages';
 
 export const load: PageServerLoad = async ({ cookies, fetch, locals, url }) => {
