+++ frontend/src/routes/list/[list_id]/edit/+page.svelte
@@ -2,8 +2,8 @@
 	import Section from '$lib/Section.svelte';
 	import { m } from '$lib/paraglide/messages.js';
 	import { enhance } from '$app/forms';
-	import { debounce } from '$lib/ui';
-	import client, { getDisplayText } from '$lib/api';
+	import { debounce, getDisplayText } from '$lib/ui';
+	import client from '$lib/api';
 	import { goto } from '$app/navigation';
 	import { draggable, droppable, type DragDropCallbacks } from '@thisux/sveltednd';
 	import { callSavingToast } from '$lib/toast';
