+++ frontend/src/routes/work/[work_id]/tags/+page.svelte
@@ -1,12 +1,13 @@
 <script lang="ts">
 	import { goto } from '$app/navigation';
-	import client, { getTagDisplaySlug } from '$lib/api.js';
+	import client from '$lib/api.js';
 	import { m } from '$lib/paraglide/messages';
 	import Section from '$lib/Section.svelte';
 	import TagsField from '$lib/TagsField.svelte';
 	import TagEditTable from '$lib/TagEditTable.svelte';
 	import GuidelineWarning from '$lib/GuidelineWarning.svelte';
 	import type { components } from '$lib/schema.js';
+	import { getTagDisplaySlug } from '$lib/ui.js';
 
 	let { data } = $props();
