+++ frontend/src/routes/work/merge/+page.svelte
@@ -4,10 +4,11 @@
 	import WorkField from '$lib/WorkField.svelte';
 	import { Rating, WorkOrigin, type components } from '$lib/schema';
 	import { enumValues, PlatformNames, RatingNames, WorkOriginNames } from '$lib/enums';
-	import client, { getDisplayText } from '$lib/api';
+	import client from '$lib/api';
 	import GuidelineWarning from '$lib/GuidelineWarning.svelte';
 	import WorkThumbnail from '$lib/WorkThumbnail.svelte';
 	import type { ComponentProps } from 'svelte';
+	import { getDisplayText } from '$lib/ui';
 
 	let work: Record<
 		'a' | 'b',
